// src/types/global.d.ts
// types/express/index.d.ts
import { JwtPayload } from "jsonwebtoken";

declare global {
  namespace Express {
    interface Request {
      user?: string | JwtPayload;
    }
  }
}


export interface PaginatedResponse<T> {
  data: T[];
  status: boolean;
  message: string;
  meta: PaginationMeta;
}

export interface ApiResponse<T> {
  status: boolean;
  message: string;
  data: T;
}

export interface PaginationMeta {
  totalCount: number;
  pageSize: number;
  currentPage: number;
  totalPages: number;
  hasNextPage: boolean;
  hasPreviousPage: boolean;
}

export type TagColorType = "success" | "error" | "warning" | undefined;